Move all exmaples into new package
