+++ examples/src/main/java/fr/epicanard/duckconfig/examples/Example.java
@@ -1,31 +1,10 @@
-package fr.epicanard.duckconfig;
+package fr.epicanard.duckconfig.examples;
 
-import fr.epicanard.duckconfig.annotations.Resource;
+import fr.epicanard.duckconfig.DuckLoader;
 import fr.epicanard.duckconfig.annotations.ResourceLocation;
 import fr.epicanard.duckconfig.annotations.ResourceWrapper;
 
 import java.util.Map;
-
-class Pat {
-  public String plop = "tutu";
-
-  @Override
-  public String toString() {
-    return String.format("[Plop:%s]", plop);
-  }
-}
-
-@Resource(value = "plop.yml", location = ResourceLocation.CLASS_PATH)
-class Plop {
-  public String name = "toto";
-  public String age = "plop";
-  public Pat pat = new Pat();
-
-  @Override
-  public String toString() {
-    return String.format("[Name:%s, Age:%s, Pat:%s]", name, age, pat);
-  }
-}
 
 public class Example {
   static void loadPlop() {
@@ -36,7 +15,7 @@ public class Example {
     System.out.println(plop);
 
     // Save
-    DuckLoader.save(plop, new ResourceWrapper(null, "plop.yml", ResourceLocation.FILE_PATH));
+    DuckLoader.save(plop, new ResourceWrapper("examples/target", "plop.yml", ResourceLocation.FILE_PATH));
 
   }
 
@@ -50,7 +29,7 @@ public class Example {
     });
 
     // Save
-    DuckLoader.save(entries, new ResourceWrapper(null, "plop_entries.yml", ResourceLocation.FILE_PATH));
+    DuckLoader.save(entries, new ResourceWrapper("examples/target", "plop_entries.yml", ResourceLocation.FILE_PATH));
   }
 
   public static void main(String[] args) {
