+++ examples/src/main/java/fr/epicanard/duckconfig/examples/Pat.java
@@ -1,0 +1,11 @@
+package fr.epicanard.duckconfig.examples;
+
+class Pat {
+  public String plop = "tutu";
+
+  @Override
+  public String toString() {
+    return String.format("[Plop:%s]", plop);
+  }
+}
+
