+++ examples/src/main/resources/entries_plop.yml
