+++ examples/src/main/resources/plop.yml
