--- src/main/java/fr/epicanard/duckconfig/annotations/Entry.java
@@ -1,12 +1,0 @@
-package fr.epicanard.duckconfig.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Entry {
-  String value() default "";
-}
